Refactor SortedKeyList into its own module
